requirements.txt
fastapi
uvicorn
pydantic
requests
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
bcrypt
sqlalchemy
